seJails y setFruits cambiado a los Level2 y 3

+++ game/backend/level/Level1.java
@@ -1,10 +1,6 @@
 package game.backend.level;
 
 import game.backend.GameState;
-import game.backend.Grid;
-import game.backend.cell.CandyGeneratorCell;
-import game.backend.cell.Cell;
-import game.backend.element.Wall;
 
 public class Level1 extends Level {
 	
@@ -18,7 +14,7 @@ public class Level1 extends Level {
 	
 	private class Level1State extends GameState {
 		private long requiredScore;
-		private long maxMoves;
+		private int maxMoves;
 		
 		public Level1State(long requiredScore, int maxMoves) {
 			this.requiredScore = requiredScore;
